updated connections to all functions. cleaned some unnecessary logic for the project atm.

+++ repository/structs.go
@@ -1,20 +1,20 @@
 package repository
 
 type Genre struct {
-	Id     int64   `json:"id"`
+	Id     int     `json:"id"`
 	Name   string  `json:"name"`
 	Movies []Movie `json:"movies"`
 }
 
 type Actor struct {
-	Id        int64   `json:"id"`
+	Id        int     `json:"id"`
 	Name      string  `json:"name"`
 	BirthDate string  `json:"birthDate"`
 	Movies    []Movie `json:"movies"`
 }
 
 type Movie struct {
-	Id          int64   `json:"id"`
+	Id          int     `json:"id"`
 	Title       string  `json:"title"`
 	ReleaseYear int     `json:"release_year"`
 	Duration    int     `json:"duration"`
